refactor: build the join graph in the unit that owns it
JoinGraphConstruction was a translation unit for a single public function,
buildJoinGraph, plus the filter-classification helper it delegates to. The
split was justified only by keeping AST and plan headers out of the cost
search, and by letting the test helper include graph construction alone --
one include line.

JoinGraph already owns the type being constructed and already included Ast,
AstNode, Variable and EnumerateCollectionNode, so folding the two functions
in costs five further includes and removes a file pair. Constructing a
JoinGraph now lives next to JoinGraph.

The cost search still sees none of those headers: it includes JoinGraph.h for
the type, not the plan-walking includes, which stay in the .cpp.
