+++ arangod/Aql/Optimizer/Rule/OptimizeJoinOrder.cpp
@@ -21,7 +21,7 @@
 
 #include "OptimizeJoinOrder.h"
 
-#include "Aql/Optimizer/Rule/OptimizeJoinOrder/JoinGraphConstruction.h"
+#include "Aql/Optimizer/Rule/OptimizeJoinOrder/JoinGraph.h"
 #include "Aql/Optimizer/Rule/OptimizeJoinOrder/JoinOrderSearch.h"
 
 #include "Aql/ExecutionPlan.h"
