+++ arangod/Aql/Optimizer/Rule/OptimizeJoinOrder/CMakeLists.txt
@@ -1,7 +1,6 @@
 target_sources(arango_aql PRIVATE
   IndexJoinStatistics.cpp
   JoinGraph.cpp
-  JoinGraphConstruction.cpp
   JoinOrderSearch.cpp
   SystemRCostEstimator.cpp
 )
