+++ arangod/Aql/Optimizer/Rule/OptimizeJoinOrder/JoinGraph.cpp
@@ -24,7 +24,13 @@
 
 #include "Aql/Ast.h"
 #include "Aql/AstNode.h"
+#include "Aql/Condition.h"
+#include "Aql/ExecutionPlan.h"
+#include "Aql/Expression.h"
+#include "Aql/ExecutionNode/CalculationNode.h"
 #include "Aql/ExecutionNode/EnumerateCollectionNode.h"
+#include "Aql/ExecutionNode/ExecutionNode.h"
+#include "Aql/ExecutionNode/FilterNode.h"
 #include "Aql/Variable.h"
 #include "Aql/types.h"
 #include "Assertions/ProdAssert.h"
@@ -199,6 +205,135 @@ auto JoinGraph::connectedComponents() const
     components.emplace_back(std::move(component));
   }
   return components;
+}
+
+namespace {
+
+// a single graph variable become node conditions, everything else is recorded
+// verbatim as a residual.
+void handleExpression(JoinGraph& graph, ExecutionPlan const* plan,
+                      AstNode const* original) {
+  Condition cond{plan->getAst()};
+  cond.andCombine(original);
+  cond.normalize();
+  AstNode const* root = cond.root();
+
+  // normalize() produces disjunctive normal form: OR( AND(...), AND(...), ...
+  // ). We can only reason about a single conjunction of predicates; anything
+  // with a real disjunction is kept as a residual for later stages.
+  if (root->type != NODE_TYPE_OPERATOR_NARY_OR || root->numMembers() != 1) {
+    graph.addResidual(original);
+    return;
+  }
+
+  auto ands = root->getMemberUnchecked(0);
+  if (ands->type != NODE_TYPE_OPERATOR_NARY_AND) {
+    graph.addResidual(original);
+    return;
+  }
+
+  auto checkIsGraphVariableAccess =
+      [&](std::optional<std::pair<Variable const*, AttributePath>> maybeAccess)
+      -> std::optional<
+          std::tuple<Variable const*, JoinGraph::Node*, AttributePath>> {
+    if (maybeAccess.has_value()) {
+      auto& [var, path] = *maybeAccess;
+      if (auto node = graph.nodeForVariable(var); node != nullptr) {
+        return std::make_tuple(var, node, std::move(path));
+      }
+    }
+    return std::nullopt;
+  };
+
+  for (size_t i = 0; i < ands->numMembers(); i++) {
+    auto predicate = ands->getMemberUnchecked(i);
+
+    // only equijoins (`==`) are understood; anything else is a residual
+    if (predicate->type != NODE_TYPE_OPERATOR_BINARY_EQ) {
+      graph.addResidual(predicate);
+      continue;
+    }
+
+    auto lhs = predicate->getMemberUnchecked(0);
+    auto maybeLhsAccess =
+        checkIsGraphVariableAccess(extractAttributeAccess(lhs));
+
+    auto rhs = predicate->getMemberUnchecked(1);
+    auto maybeRhsAccess =
+        checkIsGraphVariableAccess(extractAttributeAccess(rhs));
+
+    // canonicalize: if only one side is a graph variable, keep it on the left
+    if (maybeRhsAccess.has_value() && !maybeLhsAccess.has_value()) {
+      std::swap(maybeLhsAccess, maybeRhsAccess);
+    }
+
+    if (maybeLhsAccess.has_value() && maybeRhsAccess.has_value()) {
+      // `a.x == b.y` with both a and b in the graph -> join edge
+      [[maybe_unused]] auto& [lhsVar, lhsNode, lhsPath] =
+          maybeLhsAccess.value();
+      [[maybe_unused]] auto& [rhsVar, rhsNode, rhsPath] =
+          maybeRhsAccess.value();
+      graph.addJoinCondition(lhsVar, lhsPath, rhsVar, rhsPath);
+    } else if (maybeLhsAccess.has_value()) {
+      // `a.x == <constant>` -> constant restriction on node a
+      [[maybe_unused]] auto& [lhsVar, lhsNode, lhsPath] =
+          maybeLhsAccess.value();
+      lhsNode->conditions.emplace_back(std::move(lhsPath));
+    } else {
+      // neither side references a graph variable -> residual
+      graph.addResidual(predicate);
+    }
+  }
+}
+
+}  // namespace
+
+auto buildJoinGraph(ExecutionPlan const* plan, ExecutionNode* firstEnumeration,
+                    ExecutionNode*& next) -> JoinGraph {
+  JoinGraph graph;
+  next = nullptr;
+
+  ExecutionNode* nextNode = nullptr;
+  for (ExecutionNode* n = firstEnumeration; n != nullptr; n = nextNode) {
+    nextNode = n->getFirstParent();
+
+    switch (n->getType()) {
+      case ExecutionNode::ENUMERATE_COLLECTION: {
+        graph.addNode(ExecutionNode::castTo<EnumerateCollectionNode*>(n));
+        break;
+      }
+
+      case ExecutionNode::CALCULATION: {
+        auto* calc = ExecutionNode::castTo<CalculationNode*>(n);
+        if (!calc->expression()->isDeterministic()) {
+          graph.hasNonDeterministicCalculation = true;
+        }
+        break;
+      }
+
+      case ExecutionNode::FILTER: {
+        auto* filter = ExecutionNode::castTo<FilterNode*>(n);
+        auto* setter = plan->getVarSetBy(filter->inVariable()->id);
+        // the filter's condition is only analyzable when it is produced by a
+        // calculation; otherwise we simply leave it in place (it keeps
+        // executing, so nothing is lost) and do not model it.
+        if (setter == nullptr ||
+            setter->getType() != ExecutionNode::CALCULATION) {
+          break;
+        }
+        auto* calc = ExecutionNode::castTo<CalculationNode*>(setter);
+        handleExpression(graph, plan, calc->expression()->node());
+        break;
+      }
+
+      default:
+        // any other node type terminates the run of adjacent enumerations
+        next = n;
+        return graph;
+    }
+  }
+
+  return graph;
 }
 
 }  // namespace arangodb::aql
