--- arangod/Aql/Optimizer/Rule/OptimizeJoinOrder/JoinGraphConstruction.h
@@ -1,38 +1,0 @@
-////////////////////////////////////////////////////////////////////////////////
-/// DISCLAIMER
-///
-/// Copyright 2014-2026 ArangoDB GmbH, Cologne, Germany
-/// Copyright 2004-2014 triAGENS GmbH, Cologne, Germany
-///
-/// Licensed under the Business Source License 1.1 (the "License");
-/// you may not use this file except in compliance with the License.
-/// You may obtain a copy of the License at
-///
-///     https://github.com/arangodb/arangodb/blob/devel/LICENSE
-///
-/// Unless required by applicable law or agreed to in writing, software
-/// distributed under the License is distributed on an "AS IS" BASIS,
-/// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-/// See the License for the specific language governing permissions and
-/// limitations under the License.
-///
-/// Copyright holder is ArangoDB GmbH, Cologne, Germany
-///
-////////////////////////////////////////////////////////////////////////////////
-
-#pragma once
-
-#include "Aql/Optimizer/Rule/OptimizeJoinOrder/JoinGraph.h"
-
-namespace arangodb::aql {
-class ExecutionNode;
-class ExecutionPlan;
-
-/// @brief build the join graph for the maximal run of adjacent collection
-/// enumerations that starts at `firstEnumeration`, following parents towards
-/// the plan root. Returns the graph and, via `next`, the first node that
-/// terminated the run (nullptr if the run reached the root).
-auto buildJoinGraph(ExecutionPlan const* plan, ExecutionNode* firstEnumeration,
-                    ExecutionNode*& next) -> JoinGraph;
-
-}  // namespace arangodb::aql
