+++ tests/Aql/JoinGraphTestHelper.h
@@ -28,7 +28,6 @@
 #include "Aql/ExecutionNode/ExecutionNode.h"
 #include "Aql/ExecutionPlan.h"
 #include "Aql/Optimizer/Rule/OptimizeJoinOrder/JoinGraph.h"
-#include "Aql/Optimizer/Rule/OptimizeJoinOrder/JoinGraphConstruction.h"
 #include "Aql/Optimizer/Rule/OptimizeJoinOrder/JoinCostEstimator.h"
 #include "Aql/Optimizer/Rule/OptimizeJoinOrder/JoinStatistics.h"
 #include "Aql/Query.h"
